+++ frontend/src/components/layout/AdminSidebar.vue
@@ -1,8 +1,17 @@
 <script setup>
+import { computed } from 'vue'
 import { useRoute } from 'vue-router'
 import { adminNavigation } from '@/constants/navigation'
+import { hasAnyAdminPermission, permissionsForAdminPath } from '@/utils/admin-permissions'
+import { useAuthStore } from '@/stores/auth'
 
 const route = useRoute()
+const authStore = useAuthStore()
+const visibleNavigation = computed(() =>
+  adminNavigation.filter((item) =>
+    hasAnyAdminPermission(authStore.currentRole, permissionsForAdminPath(item.path))
+  )
+)
 </script>
 
 <template>
@@ -16,7 +25,7 @@ const route = useRoute()
     </div>
 
     <el-menu :default-active="route.meta?.navKey" class="admin-sidebar__menu" router>
-      <el-menu-item v-for="item in adminNavigation" :key="item.path" :index="item.path">
+      <el-menu-item v-for="item in visibleNavigation" :key="item.path" :index="item.path">
         {{ item.label }}
       </el-menu-item>
     </el-menu>